Remove the viewWillTransitionToSize
This animation doesn't seem to be needed anymore, and it crashes with "Application tried to present modally an active controller <SCRecordingSaveViewController>" on iOS 12 and earlier. It looks like dismissViewControllerAnimated isn't fast enough or some other nonsense. I'm removing this animation, iOS 13 and later can rotate the picker anyway and iOS 12 and earlier can suck eggs (or they can just use it without rotating)